Refactoring module into submodules. Added MatlabCompat.Image
